pyproject.toml
[tool.poetry]
name = "mycurrency"
version = "0.1.0"
description = ""
authors = ["Eduardo Nuez Wehbe <nuezwehbe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pydantic = "^2.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
